+++ src/runtime-gcc/CMakeLists.txt
@@ -18,6 +18,5 @@ target_include_directories(${NAME} PRIVATE
     ${CMAKE_CURRENT_LIST_DIR}/../../include
 )
 target_link_libraries(${NAME} PRIVATE
-    sys
+    runtime-sys
 )
-
