+++ src/runtime-gcc/hash.c
@@ -1,7 +1,7 @@
 #include "hash.h"
 #include <objc/objc.h>
 #include <string.h>
-#include <sys/sys.h>
+#include <runtime-sys/sys.h>
 
 #define HASH_TABLE_SIZE 512
 struct objc_hashitem hash_table[HASH_TABLE_SIZE + 1];
