+++ src/runtime-sys/darwin/random.c
@@ -2,7 +2,7 @@
 #include <stdbool.h>
 #include <stdint.h>
 #include <stdlib.h>
-#include <sys/sys.h>
+#include <runtime-sys/sys.h>
 #include <time.h>
 
 /**
