+++ src/runtime-sys/pico/abort.c
@@ -1,5 +1,5 @@
 #include <stdlib.h>
-#include <sys/sys.h>
+#include <runtime-sys/sys.h>
 
 void sys_abort(void) {
   sys_puts("\nHALT");
