+++ src/runtime-sys/posix/malloc.c
@@ -1,5 +1,5 @@
 #include <stdlib.h>
-#include <sys/sys.h>
+#include <runtime-sys/sys.h>
 
 /*
  ** Allocate memory
