+++ src/runtime-sys/posix/puts.c
