+++ src/runtime-sys/posix/thread.c
@@ -1,4 +1,4 @@
-#include <sys/sys.h>
+#include <runtime-sys/sys.h>
 #include <unistd.h>
 
 /**
