+++ src/tests/NXFoundation_06/CMakeLists.txt
@@ -5,4 +5,9 @@ add_executable(${NAME}
 target_link_libraries(${NAME} PRIVATE
     NXFoundation
 )
+# Ensure C11 standard is used for for-loop declarations
+set_target_properties(${NAME} PROPERTIES
+    C_STANDARD 11
+    C_EXTENSIONS ON
+)
 add_test(NAME ${NAME} COMMAND ${NAME})
