+++ src/tests/sys_01/main.c
@@ -1,4 +1,4 @@
-#include <sys/sys.h>
+#include <runtime-sys/sys.h>
 #include <tests/tests.h>
 
 int main(void) {
