+++ src/tests/sys_02/CMakeLists.txt
@@ -3,6 +3,6 @@ add_executable(${NAME}
     main.c
 )
 target_link_libraries(${NAME} PRIVATE
-    sys
+    runtime-sys
 )
 add_test(NAME ${NAME} COMMAND ${NAME})
