+++ src/tests/sys_03/main.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-#include <sys/sys.h>
+#include <runtime-sys/sys.h>
 #include <tests/tests.h>
 
 int main(void) {
